Move form and library code into their respective files
